feat: add reusable forecast chart component
