+++ src/scenes/SpotDetailsScene.tsx
@@ -1,16 +1,15 @@
-import React, { useRef, useState, useEffect, useMemo } from "react";
+import React, { useRef, useState, useEffect } from "react";
 import { ChevronLeft, Plus, Pencil, Maximize2, Trash2 } from "lucide-react";
-import { animate } from "framer-motion";
 import { Button } from "@/components/ui/button";
 import { BTN, BTN_GHOST_ICON, T_PRIMARY, T_MUTED, T_SUBTLE } from "../styles/tokens";
 import { useT } from "../i18n";
 import type { Spot, VisitHistory } from "../types";
-import { todayISO, generateForecast, formatDate } from "../utils";
+import { todayISO, formatDate } from "../utils";
 import { loadMap } from "@/services/openstreetmap";
 import type { StyleSpecification } from "maplibre-gl";
 import Logo from "@/assets/logo.png";
 import { useAppContext } from "../context/AppContext";
-import { ResponsiveContainer, LineChart, Line, XAxis, YAxis, Tooltip, ReferenceLine } from "recharts";
+import ForecastChart from "@/components/ForecastChart";
 import { ConfirmDeleteModal } from "../components/ConfirmDeleteModal";
 import { EditVisitModal } from "../components/EditVisitModal";
 
@@ -33,8 +32,6 @@ export default function SpotDetailsScene({ spot, onBack }: { spot: Spot | null; 
   const [confirmOpen, setConfirmOpen] = useState(false);
   const mapContainerRef = useRef<HTMLDivElement | null>(null);
   const mapRef = useRef<any>(null);
-  const chartRef = useRef<HTMLDivElement | null>(null);
-  const data = useMemo(() => generateForecast(state.prefs.lang), [state.prefs.lang]);
 
   useEffect(() => {
     if (!spot?.location || mapRef.current || !mapContainerRef.current) return;
@@ -76,16 +73,6 @@ export default function SpotDetailsScene({ spot, onBack }: { spot: Spot | null; 
     };
   }, [spot?.location]);
 
-  useEffect(() => {
-    const path = chartRef.current?.querySelector(
-      ".recharts-line-curve"
-    ) as SVGPathElement | null;
-    if (!path) return;
-    const length = path.getTotalLength();
-    path.style.strokeDasharray = `${length}`;
-    path.style.strokeDashoffset = `${length}`;
-    animate(path, { strokeDashoffset: 0 }, { duration: 1.2, ease: "easeInOut" });
-  }, [data]);
 
   if (!spot) return null;
   const photos = spot.photos || [];
@@ -136,39 +123,7 @@ export default function SpotDetailsScene({ spot, onBack }: { spot: Spot | null; 
         <p className={`text-xs mt-2 ${T_SUBTLE}`}>{t("La carte affiche l'historique complet avec détails.")}</p>
 
         <div className="mt-3">
-          <div className="h-56" ref={chartRef}>
-            <ResponsiveContainer width="100%" height="100%">
-              <LineChart data={data} margin={{ top: 10, right: 10, bottom: 0, left: 0 }}>
-                <XAxis
-                  dataKey="day"
-                  tick={{ fontSize: 10, fill: "hsl(var(--forest-green))" }}
-                  interval={3}
-                />
-                <YAxis
-                  domain={[0, 100]}
-                  tick={{ fontSize: 10, fill: "hsl(var(--forest-green))" }}
-                  width={28}
-                />
-                <Tooltip
-                  contentStyle={{
-                    background: "#171717",
-                    border: "1px solid #262626",
-                    borderRadius: 12,
-                    color: "#e5e5e5",
-                  }}
-                />
-                <ReferenceLine x={data[7].day} stroke="#525252" strokeDasharray="3 3" />
-                <Line
-                  type="monotone"
-                  dataKey="score"
-                  stroke="hsl(var(--fern-green))"
-                  strokeWidth={2}
-                  dot={false}
-                  isAnimationActive={false}
-                />
-              </LineChart>
-            </ResponsiveContainer>
-          </div>
+          <ForecastChart />
           <div className={`mt-2 text-xs ${T_SUBTLE}`}>Prévisions locales (démo)</div>
         </div>
